--- addons/brushclearing/functions/CfgFunctions.hpp
@@ -1,7 +1,0 @@
-class ClearBrush {
-    class ClearBrush {
-        file = "\ClearBrush\functions";
-		class clearBrush;
-		class canClearBrush;
-    };
-};
