+++ addons/main/functions/fn_canClearBrush.sqf
@@ -9,7 +9,7 @@ Check if it's destroy bush or cut grass.
 * -
 
 * Example:
-* [player] call ClearBrush_fnc_canClearBrush
+* [player] call TF47_fnc_canClearBrush
 */
 
 params ["_player"];
